Create separate module for wallpaper setting
